pyproject.toml
[project]
name = "mcp-experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "isort>=6.0.1",
    "jaydebeapi>=1.2.3",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-ollama>=0.3.6",
    "langgraph>=0.6.4",
    "mcp>=1.12.4",
    "python-dotenv>=1.1.1",
]
